multipathd.service: add dependency on initrd-cleanup.service
This is necessary in the initramfs. With this added, almost no
differences between our shipped multipathd.service and the one
shipped in dracut remain. The rest should be fixed in dracut.

It would be nice to be able to use the original service file
from multipath-tools in dracut, in particular for future changes.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>