multipathd: avoid unnecessary path read-only reloads
A mulitpath device can only be reloaded read/write when all paths are
read/write. Also, whenever a read-only scsi device is rescanned, the
scsi subsystem will first unconditionally issue a uevent with DISK_RO=0
before checking the read-only status, and if it the device is still
read-only, issuing another uevent with DISK_RO=1. These uevents cause
pointless reloads when read-only paths are rescanned. To avoid this,
first check if the path is being changed to the existing multipath
read-only state. If the state is the same, do nothing. If it's
different, check to see if all paths are read/write before changing a
multipath device from read-only to read/write. If the multipath device
read-only state is unknown, assume that it needs to be reloaded.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>