multipath-tools: Makefile.inc: add test for -D_FORTIFY_SOURCE=3
6186645 ("Fix possibility to redefine -D_FORTIFY_SOURCE macro.")
does not work as-is, because OPTFLAGS can't be used to override CPPFLAGS.
Instead, add a test for support of -D_FORTIFY_SOURCE=3, and use it
automatically if supported. The test uses similar logic as e.g.
https://sourceware.org/git/?p=elfutils.git;a=commit;h=29859f2e79ef3c650ee9712cae990c6a7f787a7d

This test works in environments with glibc 2.33 or newer. On older distributions,
-D_FORTIFY_SOURCE=3 does not cause an error, and will thus be used. In this
case, it has the same effect as -D_FORTIFY_SOURCE=2. On alpine Linux (musl
libc), -D_FORTIFY_SOURCE=3 causes an error, thus -D_FORTIFY_SOURCE=2
will be used.

Fixes: 6186645 ("Fix possibility to redefine -D_FORTIFY_SOURCE macro.")
Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Liska <mliska@suse.cz>