libmultipath: remove weak attribute for {get,put}_multipath_config
The "weak" attribute was set to make sure that these symbols are overridden by
symbols of the same name in the main executable. This is unnecessary, as the
standard ELF symbol resolution algorithm ensures this even without the "weak"
attribute.

See ELF Specification, III.2 Dynamic Linking / Shared Object Dependencies:
"The dynamic linker examines the symbol tables with a breadth-first
search. That is, it first looks at the symbol table of the executable program
itself, then at the symbol tables of the DT_NEEDED entries (in order), then at
the second level DT_NEEDED entries, and so on".

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>