multipathc: add new interactive client program
Add a new small program, multipathc, that acts as interactive
uxsock client for multipathd. multipathc is the only program
that has an interactive mode and can thus link to either libreadline
or libedit for command line history. All code depending on libreadline
is moved from uxclnt.c and cli.c to multipathc.c.

This patch breaks multipathd's interactive mode. It will be restored
in the next patch.

As multipathc doesn't link to libmultipath, it can link to libreadline
without license conflict.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>