libmultipath: avoid -Warray-bounds error with gcc 12 and musl libc
The following error is observed with gcc 12, strangely only with
MUSL libc:

In function ‘__uatomic_inc’,
     inlined from ‘lock’ at ../libmultipath/lock.h:24:2,
     inlined from ‘child’ at main.c:3523:3,
     inlined from ‘main’ at main.c:3755:11:
/usr/include/urcu/uatomic/x86.h:439:17: error: array subscript ‘struct __uatomic_dummy[0]’ is partly outside array bounds of ‘unsigned char[72]’ [-Werror=array-bounds]

The problem is that &(vecs->lock.waiters) is casted to a pointer to
struct { long[10]; } which goes beyond the "struct vectors".
We don't read or write from/to that memory, but the compiler complains either
way.

latest liburcu has a patch for it:

http://git.liburcu.org/?p=userspace-rcu.git;a=commitdiff;h=835b9ab3ca3777fe42e37e92096226ebd19ca75b

For now, just disable the warning in lock.h, using a pragma.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reported-by: Xose Vasquez Perez <xose.vasquez@gmail.com>