multipath-tools build: accept KBUILD_BUILD_TIMESTAMP from env
It would appear KBUILD_BUILD_TIMESTAMP is meant to be used the same way as in
the linux kernel build. For linux, builders are supposed to set
KBUILD_BUILD_TIMESTAMP in the environment if they want a stable timestamp.

The libddmmp makefile however tries to directly call git to get a timestamp,
which fails in a typical Debian build environment, which is not a full git
source tree.

Have libdmmp/Makefile use KBUILD_BUILD_TIMESTAMP from the environment if
available, otherwise use git as before.

mwilck: changed slightly, preserving the original logic

Signed-off-by: Chris Hofstaedtler <zeha@debian.org>
Signed-off-by: Martin Wilck <mwilck@suse.com>