libmultipath: fix max_sectors_kb on adding path
A user can change the value of max_sectors_kb on the multipath device
and its path devices.
So when a path is deleted and then re-added, its value will not be the
same as the multipath device. In that case the IOs could be mis-sized.

This was originally hit on Lustre FS on:
https://jira.whamcloud.com/browse/LU-9551

On reload, this patch re-apply max_sectors_kb value of the multipath
device on its path devices.

Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>