libmultipath: don't bother to recheck timeout
This is almost always pointless work. c->timeout has already been set.
The only reason why this recheck existed was to deal with the
possibility that the sysfs value had changed.  It is unlikely that users
will update the sysfs value to change the multipath timeout while
multipathd is running. They can alway reload multipathd if they want
it changed, anyways.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>