multipathd.service: add socket as dependency
Add multipathd.socket as a weak dependency. Before this commit, systemd knows
the start order (socket -> Service) due to the After=multipathd.socket but no
explicit dependency is made. This solves the issue if a user attempts to start
the socket after the service like so:

    systemctl stop multipathd.service multipathd.socket
    systemctl start multipathd.service multipathd.socket

An error will be noted in journalctl, and the socket will be left in an
inactive state if you query it with systemctl status multipathd.socket.