multipathd: Implement "shutdown if-idle"
Currently, socket activation doesn't work properly with
multipath-tools (see #76 for a lengthy discussion about the issue).
Due to the way multipath-tools and udev interact, it is not possible
to guarantee that the daemon will be stopped even if there are no
multipath devices/maps present in the system.

In order to workaround that limitation, this commit proposes a new
daemon command called "shutdown if-idle".  What it does is check if
any multipath maps have been found during system initialization.  If
none exist, then the daemon will shut itself down.

Closes: #76

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>